1rst part check2

+++ libft/ft_bzero.c
@@ -6,13 +6,14 @@
 /*   By: sofgonza <sofgonza@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/08 12:24:27 by sofgonza          #+#    #+#             */
-/*   Updated: 2023/03/13 12:27:51 by sofgonza         ###   ########.fr       */
+/*   Updated: 2023/03/15 15:36:59 by sofgonza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stdlib.h>
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	char			*dst;
 	unsigned int	x;
