+++ libft/ft_memchr.c
@@ -6,13 +6,14 @@
 /*   By: sofgonza <sofgonza@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 16:48:15 by sofgonza          #+#    #+#             */
-/*   Updated: 2023/03/14 17:13:34 by sofgonza         ###   ########.fr       */
+/*   Updated: 2023/03/15 15:44:21 by sofgonza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stdlib.h>
 #include "libft.h"
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char	*str;
 	unsigned int	x;
