+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: sofgonza <sofgonza@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/10 16:09:53 by sofgonza          #+#    #+#             */
-/*   Updated: 2023/03/15 15:11:07 by sofgonza         ###   ########.fr       */
+/*   Updated: 2023/03/15 15:46:28 by sofgonza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,21 +20,21 @@ int				ft_isalnum(int c);
 int				ft_isascii(int c);
 int				ft_isprint(int c);
 int				ft_strlen(const char *s);
-void			*ft_memset(void *b, int c, unsigned int len);
-void			ft_bzero(void *s, unsigned int n);
-void			*ft_memcpy(void *dest, const void *src, unsigned int n);
-void			*ft_memmove(void *dest, const void *src, unsigned int len);
+void			*ft_memset(void *b, int c, size_t len);
+void			ft_bzero(void *s, size_t n);
+void			*ft_memcpy(void *dest, const void *src, size_t n);
+void			*ft_memmove(void *dest, const void *src, size_t len);
 int				ft_toupper(int c);
 int				ft_tolower(int c);
 char			*ft_strchr(const char *s, int c);
 char			*ft_strrchr(const char *s, int c);
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size);
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size);
-int				ft_strncmp(const char *s1, const char *s2, unsigned int n);
-void			*ft_memchr(const void *s, int c, unsigned int n);
-int				ft_memcmp(const void *s1, const void *s2, unsigned int n);
+unsigned int	ft_strlcpy(char *dest, char *src, size_t size);
+unsigned int	ft_strlcat(char *dest, char *src, size_t size);
+int				ft_strncmp(const char *s1, const char *s2, size_t n);
+void			*ft_memchr(const void *s, int c, size_t n);
+int				ft_memcmp(const void *s1, const void *s2, size_t n);
 char			*ft_strnstr(const char *haystack, const char *needle,
-					unsigned int len);
+					size_t len);
 int				ft_atoi(const char *str);
 void			*ft_calloc(size_t count, size_t size);
 char			*ft_strdup(const char *s1);
